requirements.txt
Django==1.9.6
django-contact-form==1.2
django-wysiwyg-redactor==0.4.9
